test: split maintenance suites and migrate static DOM to components
Break the renderer maintenance monolith into pure grouping/commit/store owners, move settings and workbench DOM semantics into Chromium component tests, and drop Node renderToStaticMarkup as the primary settings/workbench owner while keeping pure metadata and scan-plan unit coverage.
