+++ tests/unit/renderer/settings_editor_contract.test.ts
@@ -1,135 +1,26 @@
 import { Website } from "@mdcz/shared/enums";
-import { OrderedSiteFieldEditor, parseBufferedNumberValue, ServerPathField } from "@mdcz/views/config-form";
+import { parseBufferedNumberValue } from "@mdcz/views/config-form";
 import { dedupePathAutocompleteSuggestions } from "@mdcz/views/path";
 import {
-  AdvancedSettingsFooterContent,
-  AssetDownloadsSection,
   buildAutoSaveFlatPayload,
   buildNamingPreviewConfig,
   buildSettingsBrowseState,
   buildSitePrioritySummary,
   FIELD_REGISTRY,
-  FileBehaviorTopLevelSection,
   flattenConfig,
   getSettingsSuggestions,
   mergeConfigWithFlatPayload,
   moveSitePriorityOption,
-  NamingSection,
-  NetworkTopLevelSection,
-  PathsSection,
-  ProfileCapsule,
   resolveSitePriorityOptions,
   runLatestRevisionTask,
-  SectionAnchor,
-  SettingsEditorAutosaveProvider,
-  SettingsSectionModeProvider,
-  type SettingsServices,
-  SettingsServicesProvider,
   shouldRenderFieldInSectionMode,
-  TranslateTopLevelSection,
   toggleSitePriorityOption,
   unflattenConfig,
 } from "@mdcz/views/settings";
-import { type ComponentProps, createElement, type ReactNode } from "react";
-import { renderToStaticMarkup } from "react-dom/server";
-import { type FieldValues, FormProvider, useForm } from "react-hook-form";
 import { describe, expect, it, vi } from "vitest";
-
-const noop = vi.fn();
-const testSettingsServices = {
-  browsePath: vi.fn(async () => ({})),
-  checkCookies: vi.fn(async () => ({ results: [] })),
-  decrementInFlightSaves: vi.fn(),
-  ensureWatermarkDirectory: vi.fn(async () => ({ path: "" })),
-  getInFlightSaves: vi.fn(() => 0),
-  incrementInFlightSaves: vi.fn(),
-  listCrawlerSites: vi.fn(async () => ({ sites: [] })),
-  openWatermarkDirectory: vi.fn(async () => undefined),
-  previewNaming: vi.fn(async () => ({ items: [] })),
-  probeSiteConnectivity: vi.fn(async () => ({ ok: true, message: "" })),
-  relaunchApp: vi.fn(async () => undefined),
-  resetConfig: vi.fn(async () => undefined),
-  saveConfig: vi.fn(async () => undefined),
-  testLLM: vi.fn(async () => ({ success: true, message: "" })),
-} satisfies SettingsServices;
-const createTestSettingsServices = (overrides: Partial<SettingsServices> = {}): SettingsServices => ({
-  ...testSettingsServices,
-  ...overrides,
-});
-const testSettingsNotifier = {
-  error: vi.fn(),
-  info: vi.fn(),
-  success: vi.fn(),
-};
 
 function entry(key: string) {
   return FIELD_REGISTRY.find((candidate) => candidate.key === key);
-}
-
-function FormHarness({
-  children,
-  services = testSettingsServices,
-  values = {},
-}: {
-  children?: ReactNode;
-  services?: SettingsServices;
-  values?: Record<string, unknown>;
-}) {
-  const form = useForm<FieldValues>({ defaultValues: values });
-  const flatValues = flattenConfig(values);
-
-  return createElement(
-    SettingsServicesProvider,
-    { notifier: testSettingsNotifier, services },
-    createElement(
-      FormProvider,
-      form as ComponentProps<typeof FormProvider>,
-      createElement(
-        SettingsEditorAutosaveProvider,
-        {
-          savedValues: flatValues,
-          defaultValues: flatValues,
-          defaultValuesReady: true,
-        },
-        children,
-      ),
-    ),
-  );
-}
-
-function SectionHarness({ section }: { section: "network" | "translate" | "fileBehavior" }) {
-  const values = {
-    network: {
-      proxyType: "none",
-      proxy: "",
-      useProxy: false,
-      timeout: 30,
-      retryCount: 3,
-      javdbCookie: "",
-      javbusCookie: "",
-    },
-    translate: {
-      enableTranslation: false,
-      engine: "google",
-      targetLanguage: "zh-CN",
-    },
-    behavior: {
-      successFileMove: false,
-      failedFileMove: false,
-      successFileRename: false,
-      deleteEmptyFolder: false,
-      scrapeSoftlinkPath: false,
-      saveLog: false,
-    },
-  };
-  const sectionElement =
-    section === "network"
-      ? createElement(NetworkTopLevelSection, { forceOpen: true })
-      : section === "translate"
-        ? createElement(TranslateTopLevelSection, { forceOpen: true })
-        : createElement(FileBehaviorTopLevelSection, { forceOpen: true });
-
-  return createElement(FormHarness, { values }, sectionElement);
 }
 
 describe("settings editor metadata and filtering", () => {
@@ -200,27 +91,6 @@ describe("settings editor metadata and filtering", () => {
       paths: { defaultScanExcludeDirs: ["failed_22", "/archive/output"] },
       aggregation: { fieldPriorities: { durationSeconds: ["dmm_tv", "avbase"] } },
     });
-  });
-
-  it("keeps scan exclusion directories in the paths settings surface", () => {
-    const html = renderToStaticMarkup(
-      createElement(
-        FormHarness,
-        {
-          values: {
-            paths: {
-              failedOutputFolder: "failed",
-              defaultScanExcludeDirs: ["E:/Output", "failed_22"],
-            },
-          },
-        },
-        createElement(PathsSection),
-      ),
-    );
-
-    expect(html).toContain("E:/Output");
-    expect(html).toContain("failed_22");
-    expect(html.match(/aria-autocomplete="list"/g)?.length).toBeGreaterThanOrEqual(2);
   });
 
   it("applies PRD visibility rules for normal, advanced, modified, group, and deep-link browsing", () => {
@@ -597,246 +467,5 @@ describe("settings editor render contracts", () => {
       { label: "Drive G", path: "G:/" },
       { label: "Movies", path: "G:/Movies/" },
     ]);
-  });
-
-  it("keeps OrderedSiteFieldEditor simple mode stable while rendering grouped row details", () => {
-    const simpleHtml = renderToStaticMarkup(
-      createElement(
-        FormHarness,
-        null,
-        createElement(OrderedSiteFieldEditor, {
-          value: ["javdb", "dmm"],
-          options: ["dmm", "javdb", "avbase"],
-          onChange: noop,
-        }),
-      ),
-    );
-
-    expect(simpleHtml).toContain("已启用 2/3");
-    expect(simpleHtml).toContain(">avbase<");
-    expect(simpleHtml.indexOf(">javdb<")).toBeLessThan(simpleHtml.indexOf(">dmm<"));
-
-    const groupedHtml = renderToStaticMarkup(
-      createElement(
-        FormHarness,
-        null,
-        createElement(OrderedSiteFieldEditor, {
-          value: ["dmm"],
-          options: ["dmm", "dmm_tv", "javdb"],
-          onChange: noop,
-          rows: [
-            {
-              id: "dmm_family",
-              label: "DMM/FANZA 系",
-              description: "官方售卖与配信源",
-              checkboxState: "indeterminate" as const,
-              trailingControl: createElement("span", null, "日文"),
-              chips: [
-                { label: "dmm / dmm_tv", monospace: true, variant: "outline" as const },
-                { label: "已启用 1/2", variant: "soft" as const },
-              ],
-            },
-          ],
-          selectedCount: 1,
-          totalCount: 1,
-          onToggleRow: noop,
-          onMoveRow: noop,
-          onSelectAll: noop,
-          onClearAll: noop,
-        }),
-      ),
-    );
-
-    expect(groupedHtml).toContain("DMM/FANZA 系");
-    expect(groupedHtml).toContain("官方售卖与配信源");
-    expect(groupedHtml).toContain("日文");
-    expect(groupedHtml).toContain("dmm / dmm_tv");
-    expect(groupedHtml).toContain("已启用 1/2");
-    expect(groupedHtml).toMatch(/data-state="indeterminate"|aria-checked="mixed"/);
-  });
-
-  it("renders loading profile identity without the old default-profile fallback", () => {
-    const html = renderToStaticMarkup(
-      createElement(ProfileCapsule, {
-        profiles: [],
-        activeProfile: null,
-        isLoading: true,
-        onSwitchProfile: noop,
-        onCreateProfile: noop,
-        onDeleteProfile: noop,
-        onResetConfig: noop,
-        onExportProfile: noop,
-        onImportProfile: noop,
-      }),
-    );
-
-    expect(html).toContain("aria-busy");
-    expect(html).not.toContain("默认配置");
-  });
-
-  it("defers heavy section bodies unless a section is force-opened", () => {
-    const deferredProps = {
-      id: "custom",
-      label: "Custom",
-      title: "Custom",
-      deferContent: true,
-      estimatedContentHeight: 320,
-    } satisfies Omit<ComponentProps<typeof SectionAnchor>, "children">;
-    const forceOpenProps = {
-      id: "custom-force",
-      label: "Custom Force",
-      title: "Custom Force",
-      deferContent: true,
-      forceOpen: true,
-      estimatedContentHeight: 320,
-    } satisfies Omit<ComponentProps<typeof SectionAnchor>, "children">;
-    const deferredHtml = renderToStaticMarkup(
-      createElement(
-        SectionAnchor,
-        deferredProps as ComponentProps<typeof SectionAnchor>,
-        createElement("div", null, "Deferred content"),
-      ),
-    );
-    const forceOpenHtml = renderToStaticMarkup(
-      createElement(
-        SectionAnchor,
-        forceOpenProps as ComponentProps<typeof SectionAnchor>,
-        createElement("div", null, "Force-open content"),
-      ),
-    );
-
-    expect(deferredHtml).toContain('data-deferred-placeholder="true"');
-    expect(deferredHtml).not.toContain("Deferred content");
-    expect(forceOpenHtml).toContain("Force-open content");
-    expect(forceOpenHtml).not.toContain('data-deferred-placeholder="true"');
-  });
-
-  it("hides the advanced settings footer while search filters are active", () => {
-    const filteredHtml = renderToStaticMarkup(
-      createElement(AdvancedSettingsFooterContent, {
-        hasActiveFilters: true,
-        isAdvancedVisible: false,
-        onToggleShowAdvanced: noop,
-      }),
-    );
-    const browseHtml = renderToStaticMarkup(
-      createElement(AdvancedSettingsFooterContent, {
-        hasActiveFilters: false,
-        isAdvancedVisible: false,
-        onToggleShowAdvanced: noop,
-      }),
-    );
-
-    expect(filteredHtml).not.toContain("显示高级设置");
-    expect(browseHtml).toContain("显示高级设置");
-  });
-
-  it("hides unsupported path browse buttons while keeping server path suggestions", () => {
-    const services = createTestSettingsServices({
-      getPathSuggestions: () => [
-        { label: "Movies", path: "E:/Movies" },
-        { label: "Output", path: "E:/Output" },
-      ],
-      isServer: true,
-    });
-    const html = renderToStaticMarkup(
-      createElement(
-        FormHarness,
-        { services, values: { paths: { mediaPath: "" } } },
-        createElement(ServerPathField, {
-          field: {
-            name: "paths.mediaPath",
-            onBlur: noop,
-            onChange: noop,
-            ref: noop,
-            value: "",
-          },
-        }),
-      ),
-    );
-
-    expect(html).not.toContain("<button");
-    expect(html).not.toContain("<datalist");
-    expect(html).toContain('aria-autocomplete="list"');
-  });
-
-  it("renders the PRD split sections and keeps advanced-only content out of public rows", () => {
-    const networkHtml = renderToStaticMarkup(createElement(SectionHarness, { section: "network" }));
-    const translateHtml = renderToStaticMarkup(createElement(SectionHarness, { section: "translate" }));
-    const behaviorHtml = renderToStaticMarkup(createElement(SectionHarness, { section: "fileBehavior" }));
-    const namingHtml = renderToStaticMarkup(
-      createElement(
-        FormHarness,
-        { values: { naming: { folderTemplate: "{actor}/{number}", fileTemplate: "{number}" } } },
-        createElement(NamingSection),
-      ),
-    );
-    const advancedDownloadHtml = renderToStaticMarkup(
-      createElement(
-        FormHarness,
-        { values: { download: { downloadPoster: true, sceneImageConcurrency: 4 } } },
-        createElement(SettingsSectionModeProvider, { mode: "advanced" }, createElement(AssetDownloadsSection)),
-      ),
-    );
-
-    expect(networkHtml).toContain("网络连接");
-    expect(networkHtml).toContain("代理类型");
-    expect(networkHtml).toContain("JavDB Cookie");
-    expect(translateHtml).toContain("翻译服务");
-    expect(translateHtml).toContain("翻译引擎");
-    expect(behaviorHtml).toContain("文件行为");
-    expect(behaviorHtml).toContain("成功后移动文件");
-    expect(namingHtml).toContain('aria-label="查看文件夹模板占位符"');
-    expect(namingHtml).toContain('aria-label="查看文件名模板占位符"');
-    expect(namingHtml).not.toContain("可用占位符：{actor}");
-    expect(advancedDownloadHtml).toContain("剧照下载并发");
-    expect(advancedDownloadHtml).not.toContain("下载海报");
-  });
-
-  it("shows poster badge controls only when poster downloads stay enabled and badge processing is on", () => {
-    const posterDisabledHtml = renderToStaticMarkup(
-      createElement(
-        FormHarness,
-        { values: { download: { downloadPoster: false, tagBadges: true } } },
-        createElement(AssetDownloadsSection),
-      ),
-    );
-    const hiddenHtml = renderToStaticMarkup(
-      createElement(
-        FormHarness,
-        { values: { download: { downloadPoster: true, tagBadges: false } } },
-        createElement(AssetDownloadsSection),
-      ),
-    );
-    const visibleHtml = renderToStaticMarkup(
-      createElement(
-        FormHarness,
-        {
-          values: {
-            download: {
-              downloadPoster: true,
-              tagBadges: true,
-              tagBadgeTypes: ["subtitle", "leak"],
-              tagBadgePosition: "topRight",
-            },
-          },
-        },
-        createElement(AssetDownloadsSection),
-      ),
-    );
-
-    expect(posterDisabledHtml).not.toContain("为封面添加标签角标");
-    expect(posterDisabledHtml).not.toContain("角标类型");
-    expect(posterDisabledHtml).not.toContain("角标位置");
-    expect(posterDisabledHtml).not.toContain("覆盖角标图片");
-    expect(hiddenHtml).toContain("为封面添加标签角标");
-    expect(hiddenHtml).not.toContain("角标类型");
-    expect(hiddenHtml).not.toContain("角标位置");
-    expect(hiddenHtml).not.toContain("覆盖角标图片");
-    expect(visibleHtml).toContain("角标类型");
-    expect(visibleHtml).toContain("角标位置");
-    expect(visibleHtml).toContain("覆盖角标图片");
-    expect(visibleHtml).toContain("中字");
-    expect(visibleHtml).toContain("流出");
   });
 });
